가입인사랑 프로젝트 갤러리 틀 적용

+++ src/main/java/com/artineer/artineer_renewal/service/ProjectService.java
@@ -70,6 +70,11 @@ public class ProjectService {
             if (fileName!=null) fileNames.add(fileName);
         }
 
+        // 파일이 하나도 없다면 기본 파일 "no_image.jpg" 추가
+        if (fileNames.isEmpty() || fileNames.stream().noneMatch(fileName -> fileName.matches(".*\\.(jpg|jpeg|png)$"))) {
+            fileNames.add("no_image.jpg");
+        }
+
         String fileNameString = String.join(",", fileNames);
 
         // System.out.println(fileNames);
@@ -149,14 +154,18 @@ public class ProjectService {
             Path filePath = Paths.get(fileDir + fileName);
             File file = filePath.toFile();
 
-            if(file.exists()) {
-                if(file.delete()) {
-                    System.out.println("파일 삭제 완료" + fileName);
+            if (!fileName.equals("no_image.jpg")) {
+                if (file.exists()) {
+                    if (file.delete()) {
+                        System.out.println("파일 삭제 완료: " + fileName);
+                    } else {
+                        System.out.println("파일 삭제 실패: " + fileName);
+                    }
                 } else {
-                    System.out.println("파일 삭제 실패" + fileName);
+                    System.out.println("파일을 찾을 수 없습니다: " + fileName);
                 }
             } else {
-                System.out.println("파일을 찾을 수 없습니다." + fileName);
+                System.out.println("no_image.jpg는 삭제할 수 없습니다.");
             }
 
 //            파일 없는 게시판 삭제 오류 때문에 주석처리(파일이 있을 때는 삭제 오류 없음)
